refactor: group validate command files into validate-command/ subfolder
Move validate-command.ts, cli-utils.ts, and the test into
src/validate-command/ with an index.ts barrel. validate-changelog.ts
stays at src/ since it's part of the public API.
