+++ src/validate-command/cli-utils.ts
