+++ src/validate-command/index.ts
@@ -1,0 +1,2 @@
+export { validate, type ValidateOptions } from './validate';
+export { error } from './cli-utils';
